feat: se le agrego modificaciones a la entidad de Contacto en el backend

+++ src/main/java/com/thekingmoss/domain/repository/IContactoRepository.java
@@ -3,5 +3,5 @@ package com.thekingmoss.domain.repository;
 import com.thekingmoss.domain.entity.Contacto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IContactoRepository extends JpaRepository<Contacto, Integer> {
+public interface IContactoRepository extends JpaRepository<Contacto, Long> {
 }
